+++ mrmailer/dto/message.go
@@ -12,7 +12,7 @@ type (
 	Message struct {
 		Channel       string
 		SendAfter     time.Time
-		RetryAttempts uint32
+		RetryAttempts int16
 		Data          MessageData
 	}
